Create close-open-files.ps1
The purpose of this script is to close all open files under a specified share path on a windows server. It's a rough script as many errors will be received, but the overall function does work.

This was used in conjunction with updating DFS namespace pointers. Forcing the files to close forced users to retry and end up in the desired destination.